Move BufferBuilder optimizations into wrapper class, nothing works
