+++ src/main/java/me/jellysquid/mods/sodium/client/render/vertex/VertexFormatDescriptionImpl.java
@@ -6,6 +6,7 @@ import net.caffeinemc.mods.sodium.api.vertex.format.VertexFormatDescription;
 import net.minecraft.client.render.VertexFormat;
 
 import java.util.Arrays;
+import java.util.EnumSet;
 import java.util.NoSuchElementException;
 
 public class VertexFormatDescriptionImpl implements VertexFormatDescription {
@@ -14,11 +15,28 @@ public class VertexFormatDescriptionImpl implements VertexFormatDescription {
 
     private final int[] offsets;
 
+    private final boolean isSimple;
+
     public VertexFormatDescriptionImpl(VertexFormat format, int id) {
         this.id = id;
         this.stride = format.getVertexSizeByte();
 
         this.offsets = getOffsets(format);
+        this.isSimple = checkSimple(format);
+    }
+
+    private static boolean checkSimple(VertexFormat format) {
+        EnumSet<CommonVertexAttribute> attributeSet = EnumSet.noneOf(CommonVertexAttribute.class);
+        var elementList = format.getElements();
+
+        for (int elementIndex = 0; elementIndex < elementList.size(); elementIndex++) {
+            var element = elementList.get(elementIndex);
+            if (!attributeSet.add(CommonVertexAttribute.getCommonType(element))) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     public static int[] getOffsets(VertexFormat format) {
@@ -65,5 +83,10 @@ public class VertexFormatDescriptionImpl implements VertexFormatDescription {
     @Override
     public int stride() {
         return this.stride;
+    }
+
+    @Override
+    public boolean isSimpleFormat() {
+        return this.isSimple;
     }
 }
