+++ src/main/resources/sodium.mixins.json
@@ -22,6 +22,7 @@
     "core.render.immediate.consumer.OutlineVertexConsumerMixin",
     "core.render.immediate.consumer.OverlayVertexConsumerMixin",
     "core.render.immediate.consumer.SpriteTexturedVertexConsumerMixin",
+    "core.render.immediate.consumer.VertexConsumerProviderImmediateMixin",
     "core.render.immediate.consumer.VertexConsumersMixin$DualMixin",
     "core.render.immediate.consumer.VertexConsumersMixin$UnionMixin",
     "core.render.world.BufferBuilderStorageMixin",
